refactor: move to assignment.ts
